app/requirements.txt
fastapi==0.108.0
uvicorn==0.25.0
pydantic==2.5.2
web3==6.11.1
python-dotenv==1.0.0
pytest==7.4.3
pytest-asyncio==0.23.2
httpx==0.25.2
python-multipart==0.0.6
websockets==12.0